+++ src/components/Buttons/CommentLikeButton.tsx
@@ -2,10 +2,7 @@ import React, { SetStateAction, useState } from 'react';
 import { css } from '@emotion/react';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 import { Comment, CommentsPagination, Replies } from '@/types/comment';
-import {
-  fetchAddLikeComment,
-  fetchDeleteLikeComment,
-} from '@/api/comments/comments';
+import { fetchAddLikeComment, fetchDeleteLikeComment } from '@/api/comments';
 import { useMemberQuery } from '@/hooks/api/useMemberQuery';
 import { useParseJwt } from '@/hooks/common/useParseJwt';
 import { useNewSelector } from '@/store';
