+++ src/components/Buttons/HeartButton.tsx
@@ -9,7 +9,7 @@ import { selectAccessToken } from '@/store/auth';
 import { useNewSelector } from '@/store';
 import { useParseJwt } from '@/hooks/common/useParseJwt';
 import { useMemberQuery } from '@/hooks/api/useMemberQuery';
-import { fetchAddLike, fetchDeleteLike } from '../../api/posts/posts';
+import { fetchAddLike, fetchDeleteLike } from '../../api/posts';
 import { Post, PostWithPagenation } from '../../types/post';
 import { DisabledHeart, Hearts } from '../../assets';
 import { pulsate } from '../../styles/keyframes';
