+++ src/components/common/BalanceOptionCard/BalanceOptionCard.tsx
@@ -6,9 +6,9 @@ import { useMemberQuery } from '@/hooks/api/useMemberQuery';
 import { useParseJwt } from '@/hooks/common/useParseJwt';
 import { useNewSelector } from '@/store';
 import { selectAccessToken } from '@/store/auth';
-import { voteBalanceOption } from '@/api/votes/vote';
+import { voteBalanceOption } from '@/api/vote';
 import { VoteInfo, BalanceOption } from '@/types/post';
-import { getVoteCount } from '@/api/posts/posts';
+import { getVoteCount } from '@/api/posts';
 import { isFinished } from '@/utils/date';
 import { Check, Winner } from '../../../assets';
 import DefaultImage from '../../../../public/defaultImage.png';
