+++ src/components/common/Modal/DeleteCommentModal/DeleteCommentModal.tsx
@@ -1,7 +1,7 @@
 import React, { SetStateAction, useEffect, useRef } from 'react';
 import Button from '@/components/common/Button/Button';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
-import { deleteComment } from '@/api/comments/comments';
+import { deleteComment } from '@/api/comments';
 import {
   deleteCommentModalWrapper,
   titleWrapper,
