+++ src/hooks/api/useMemberPwMutation.ts
@@ -1,4 +1,4 @@
-import { putMemberPw } from '@/api/member/member';
+import { putMemberPw } from '@/api/member';
 import { useMutation } from '@tanstack/react-query';
 
 export const useMemberPwMutation = () => {
