+++ src/hooks/api/useMemberQuery.ts
@@ -1,4 +1,4 @@
-import { getMember } from '@/api/member/member';
+import { getMember } from '@/api/member';
 import { Member } from '@/types/member';
 import { useQuery } from '@tanstack/react-query';
