+++ src/pages/PostPage/BalanceOptionCardsSection/ResultSection/ResultSection.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { useQuery } from '@tanstack/react-query';
-import { getVoteCount } from '@/api/posts/posts';
+import { getVoteCount } from '@/api/posts';
 import { VoteInfo } from '@/types/post';
 import { Fight } from '@/assets';
 import { getPercent } from '@/utils/calculator';
