+++ src/pages/PostPage/CommentsSection/CommentsSection.tsx
@@ -1,6 +1,6 @@
 import React, { SetStateAction, useState } from 'react';
 import { useQuery } from '@tanstack/react-query';
-import { getComments } from '@/api/comments/comments';
+import { getComments } from '@/api/comments';
 import { Comment } from '@/types/comment';
 import InputNewComment from '@/components/common/InputComment/InputNewComment/InputNewComment';
 import { useCreateCommentForm } from '@/hooks/comment/useCreateCommentForm';
